Move implementations into modular sources
